Unsigned number can't be negated in the VisualStudio compiler. Fixed a few earlier review comments that had slipped through.
git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@1060 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/regexp-macro-assembler-ia32.h
@@ -178,7 +178,7 @@ class RegExpMacroAssemblerIA32: public RegExpMacroAssembler {
   inline Register backtrack_stackpointer() { return ecx; }
 
   // Byte size of chars in the string to match (decided by the Mode argument)
-  inline size_t char_size() { return static_cast<size_t>(mode_); }
+  inline int char_size() { return static_cast<int>(mode_); }
 
   // Equivalent to a conditional branch to the label, unless the label
   // is NULL, in which case it is a conditional Backtrack.
@@ -225,7 +225,7 @@ class RegExpMacroAssemblerIA32: public RegExpMacroAssembler {
   // constants.
   ByteArrayProvider constants_;
 
-  // Which mode to generate code for (ASCII or UTF16).
+  // Which mode to generate code for (ASCII or UC16).
   Mode mode_;
 
   // One greater than maximal register index actually used.
